refactor

+++ workstation/setup/docker.sh
@@ -18,7 +18,7 @@ fi
 ESSENTIALS_HASH=$(echo "$KIRA_HOME-" | md5)
 SETUP_CHECK="$KIRA_SETUP/docker-1-$ESSENTIALS_HASH"
 SETUP_CHECK_REBOOT="$SETUP_CHECK-reboot"
-if [ ! -f "$SETUP_CHECK" ] || [ "${VERSION,,}" == "error" ] || (! $(isServiceActive "docker")) ; then
+if [ ! -f "$SETUP_CHECK" ] || [ "${VERSION,,}" == "error" ] || ( [ "${IS_WSL,,}" != "true" ]  && (! $(isServiceActive "docker")) ) ; then
     echoInfo "INFO: Attempting to remove old docker..."
     docker system prune -f || echoWarn "WARNING: failed to prune docker system"
     $KIRA_COMMON/docker-stop.sh || echoWarn "WARNING: Failed to stop docker servce"
